+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Add first documentation.
 * Add first cluster scope and actions.
 * New action to verifiy if host network pods matches with k8scloudconfig on master nodes
+* New action to verifiy if host network pods matches with k8scloudconfig on worker nodes
 
 [Unreleased]: https://github.com/giantswarm/awscnfm/compare/v12.1.0...HEAD
 [12.1.0]: https://github.com/giantswarm/awscnfm/releases/tag/v12.1.0